src/use-cases/pets/get-pet.ts
import { Pet } from '@prisma/client'
import { IPetsRepository } from '../../repositories/pets-repository'
import { ResourceNotFoundError } from '../errors/resource.not-found'

interface GetPetUseCaseRequest {
  petId: string
}

interface GetPetUseCaseResponse {
  pet: Pet
}

export class GetPetUseCase {
  constructor(private petsRepository: IPetsRepository) {}

  async execute({
    petId,
  }: GetPetUseCaseRequest): Promise<GetPetUseCaseResponse> {
    const pet = await this.petsRepository.findById(petId)

    if (!pet) {
      throw new ResourceNotFoundError()
    }

    return { pet }
  }
}
